+++ src/redturtle/prenotazioni/content/prenotazioni_folder.py
@@ -1,4 +1,5 @@
 # -*- coding: utf-8 -*-
+import re
 from typing import Generator
 
 from collective.z3cform.datagridfield.datagridfield import DataGridFieldFactory
@@ -209,6 +210,19 @@ def notify_on_refuse_message_default_factory(context):
     )
 
 
+def holidays_constraint(value: list):
+    if not value:
+        return True
+
+    pattern = re.compile(r"^(0[1-9]|[12][0-9]|3[01])/(0[1-9]|1[0-2])/(?:\d{4}|\*)$")
+
+    for i in value:
+        if not bool(re.match(pattern, i)):
+            return False
+
+    return True
+
+
 class IPrenotazioniFolder(model.Schema):
     """Marker interface and Dexterity Python Schema for PrenotazioniFolder"""
 
@@ -402,6 +416,7 @@ class IPrenotazioniFolder(model.Schema):
             "25/12/*",
             "26/12/*",
         ],
+        constraint=holidays_constraint,
     )
 
     futureDays = schema.Int(